2nd page gets cleared without errors

+++ skeleton/TST_RAST.C
@@ -34,7 +34,7 @@ void test_plot_bitmap_32(UINT32 *base, const UINT32 *bitmap, const unsigned int 
 int main()
 {
 	int i;
-    UINT32 *base  = Physbase();
+    UINT8 *base  = Physbase();
 	
 	const UINT8* glyph_A = GLYPH_START('A');
